Add files via upload
Based on users' prior actions on the network, an Instagram recommendation system is made to make content recommendations to them. The accounts they follow, the posts they interact with, and the hashtags they utilise are a few examples of this. The system analyses user data using machine learning techniques to generate recommendations that are likely to be of interest to each unique user.